+++ .changeset/hot-cats-marry.md
@@ -1,0 +1,5 @@
+---
+'@leafygreen-ui/marketing-modal': minor
+---
+
+`title` prop is now typed as ReactNode to allow customizations
