+++ packages/marketing-modal/src/BlobSVG/BlobSVG.styles.ts
