+++ packages/marketing-modal/src/BlobSVG/BlobSVG.types.ts
@@ -1,6 +1,6 @@
 import { Theme } from '@leafygreen-ui/lib';
 
-import { BlobPosition } from '../MarketingModal/types';
+import { BlobPosition } from '../MarketingModal/MarketingModal.types';
 
 export interface BlobSVGProps {
   blobPosition: BlobPosition;
