+++ packages/marketing-modal/src/MarketingModal/MarketingModal.styles.ts
